pythonProject/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==40.0.2
discord.py==2.2.2
distro==1.8.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
jsonschema==3.2.0
load-dotenv==0.1.0
macholib==1.16.2
multidict==6.0.4
openai==0.26.3
packaging==19.2
paramiko==3.1.0
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dotenv==0.21.1
python-telegram-bot==13.15
PyYAML==5.4.1
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
texttable==1.6.7
tqdm==4.64.1
urllib3==1.26.14
websocket-client==0.59.0
yarl==1.8.2
